Reorganize casters
