+++ app/Support/OutputResolver.php
@@ -7,11 +7,11 @@ namespace Syntatis\WP\Option\Support;
 use Syntatis\WP\Option\Contracts\Castable;
 use Syntatis\WP\Option\Contracts\Resolvable;
 use Syntatis\WP\Option\Option;
-use Syntatis\WP\Option\TypeCasters\TypeArray;
-use Syntatis\WP\Option\TypeCasters\TypeBoolean;
-use Syntatis\WP\Option\TypeCasters\TypeFloat;
-use Syntatis\WP\Option\TypeCasters\TypeInteger;
-use Syntatis\WP\Option\TypeCasters\TypeString;
+use Syntatis\WP\Option\Casters\TypeArray;
+use Syntatis\WP\Option\Casters\TypeBoolean;
+use Syntatis\WP\Option\Casters\TypeFloat;
+use Syntatis\WP\Option\Casters\TypeInteger;
+use Syntatis\WP\Option\Casters\TypeString;
 
 use function array_key_exists;
 use function is_array;
